Changed logic to use env var

+++ lib/cwmp.ts
@@ -806,6 +806,7 @@ async function endSession(sessionContext: SessionContext): Promise<void> {
   }
 
   const promises = [];
+  const noFaultCache = config.get("CWMP_NO_FAULTS_CACHE");
 
   promises.push(
     db.saveDevice(
@@ -819,17 +820,17 @@ async function endSession(sessionContext: SessionContext): Promise<void> {
   if (sessionContext.operationsTouched) {
     for (const k of Object.keys(sessionContext.operationsTouched)) {
       saveCache = true;
-      // if (sessionContext.operations[k]) {
-      //   promises.push(
-      //     db.saveOperation(
-      //       sessionContext.deviceId,
-      //       k,
-      //       sessionContext.operations[k]
-      //     )
-      //   );
-      // } else {
+      if (sessionContext.operations[k] && !noFaultCache) {
+        promises.push(
+          db.saveOperation(
+            sessionContext.deviceId,
+            k,
+            sessionContext.operations[k]
+          )
+        );
+      } else {
         promises.push(db.deleteOperation(sessionContext.deviceId, k));
-      // }
+      }
     }
   }
 
@@ -854,16 +855,16 @@ async function endSession(sessionContext: SessionContext): Promise<void> {
     }
   }
 
-  if (saveCache) {
-    // promises.push(
-    //   cacheDueTasksAndFaultsAndOperations(
-    //     sessionContext.deviceId,
-    //     sessionContext.tasks,
-    //     sessionContext.faults,
-    //     sessionContext.operations,
-    //     sessionContext.cacheUntil
-    //   )
-    // );
+  if (saveCache && !noFaultCache) {
+    promises.push(
+      cacheDueTasksAndFaultsAndOperations(
+        sessionContext.deviceId,
+        sessionContext.tasks,
+        sessionContext.faults,
+        sessionContext.operations,
+        sessionContext.cacheUntil
+      )
+    );
   }
 
   await Promise.all(promises);
